Polishing, S01E07
- Use `int` instead of `int64_t` for small loops

+++ src/download/util.cpp
@@ -12,7 +12,7 @@ std::array<uint8_t, 4> encode_integer(int64_t n) {
   auto encode_me = static_cast<uint32_t>(n);
 
   std::array<uint8_t, 4> result{};
-  for (int64_t i = 3; i >= 0; i--) {
+  for (int i = 3; i >= 0; i--) {
     result[i] = encode_me & 0xFF;
     encode_me >>= 8;
   }
@@ -21,7 +21,7 @@ std::array<uint8_t, 4> encode_integer(int64_t n) {
 
 int64_t decode_integer(const std::array<uint8_t, 4>& buf) {
   uint32_t result = 0;
-  for (int64_t i = 0; i <= 3; i++) {
+  for (int i = 0; i <= 3; i++) {
     result |= buf[i] << (8 * (3 - i));
   }
   // An uint32_t always fits inside a int64_t
